Repo Info
Some extra starter files that I usually add to new repos of mine.
Found that you have to define which files you don't want to be bundled into the executable. Had to make some manual changes to the `Package.swift` file, which is says specifically not to do, so we'll see what happens there haha. At least the warning about the repo files is gone now! Too bad they have to hide the `Package.swift` file in Playgrounds.

Some Apple docs came to the rescue this time, the `.executableTarget()` method and it's parameters:
https://docs.swift.org/package-manager/PackageDescription/PackageDescription.html