Initial File Reading
Worked on extending the file reading functionality, and now I almost have it so you can read the contents of the file (in the console as of now)! Next I am going to figure out how you can get the result and forward it to the UI, as I'd like it to show up in the `TextEditor()`.

A few more links helped out again today:
https://www.youtube.com/watch?app=desktop&v=q8y_eRVfpMA
https://stackoverflow.com/questions/64149342/how-to-add-file-picker-to-the-app-on-ios-14-and-lower
https://stackoverflow.com/questions/31352593/how-to-print-details-of-a-catch-all-exception-in-swift