File Reading on macOS
Figured out getting the `UIDocumentPicker` to work in macOS! Haven't been able to get it to work like the iOS version up until now, and it was all because I just had to add file system access as a capability in the app options. It's either a higly technical issue causing you pain, or the other 97% of the time, it's a simple issue that you just haven't found yet XD
One thing I am going to investigate next is that the Sheet view doesn't make too much sense in macOS, since the picker opens in it's own window, so I'll have to see if I can make it optional to whether it is on mobile, or desktop where multiple windows are available.
Also, I'm not totally sure if `.swiftpm` is supposed to go into your `.gitignore`, but I'm leaving it in the repo for the meantime. Better to have extra info than to lose what you might actually need :)

Thanks again, internet!
(I want to use that somewhere else, that's a nice phrase haha)
https://stackoverflow.com/questions/56645819/how-to-open-file-dialog-with-swiftui-on-platform-uikit-for-mac/57398756#57398756
https://appventure.me/guides/catalyst/how/open_save_export_import.html