Next File Reading
Actually got the file reading to, well, read! Pretty cool, haha.

This first link was the one that ended up working:
https://stackoverflow.com/questions/34636150/no-permission-to-view-document-passed-back-from-ios-document-provider-on-open-op
https://stackoverflow.com/questions/67438641/how-to-access-a-security-scoped-resource-from-a-string-path-in-swift

This is one that I found right after I got the solution to work, and it appears to be the new built-in SwiftUI view for opening files. It must be pretty new because none of the other articles even mentioned it as a solution or reference (`.fileImporter`). I'm gonna try to see if that's another option next, I just wanted to make sure to commit this so I can come back to this working version later if the other one isn't a solution.
https://stackoverflow.com/questions/70343113/swiftui-display-file-url-from-array-after-user-picks-file-using-documentpicker