File Loading: fileImporter
Yet another iteration of testing how to open files, and this time I used the `.fileImporter()` method that I found in the article at the end of the last update. It works really well! Now that this is fairly consistent, I'm going to move on to using different UI components, as I haven't really expanded on that with the app yet. I kind of want to make a lightweight text editor, but I've done that before already, so I'm gonna just keep working until I find what this project will turn into. Not quite sure yet!

https://developer.apple.com/documentation/swiftui/image/fileimporter(ispresented:allowedcontenttypes:allowsmultipleselection:oncompletion:)
https://stackoverflow.com/questions/64974785/swiftui-loading-images-with-fileimporter